pyproject.toml
[tool.poetry]
name = "sudoku-solver"
version = "0.1.0"
description = ""
authors = ["andraspalasti <andraspalasti2@gmail.com>"]
readme = "README.md"
packages = [{include = "ml"}]

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.0.1"
pillow = "^10.0.0"
torchvision = "^0.15.2"
tqdm = "^4.65.0"
albumentations = "^1.3.1"
numpy = "^1.25.0"
pandas = "^2.0.3"
opencv-python = "^4.8.0.74"
onnx = "^1.14.0"


[tool.poetry.group.dev.dependencies]
matplotlib = "^3.7.2"
ipykernel = "^6.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
